Apply further renaming suggestions
