+++ src/DSE.Open.Numerics/ReadOnlyDataFrameExtensions.cs
@@ -6,20 +6,28 @@ namespace DSE.Open.Numerics;
 public static class ReadOnlyDataFrameExtensions
 {
     /// <summary>
-    /// Gets the specified typed series from the <paramref name="dataFrame"/>, if present. Otherwise, throws.
+    /// Gets the series with the specified name from the data frame.
     /// </summary>
-    /// <param name="dataFrame">The data-frame.</param>
-    /// <param name="name">The name of the series.</param>
-    /// <typeparam name="T">The expected <em>element</em> type of the series.</typeparam>
-    /// <exception cref="KeyNotFoundException">The <paramref name="dataFrame"/> does not contain a series named <paramref name="name"/>.</exception>
+    /// <exception cref="KeyNotFoundException">Thrown when the series with the specified name does not exist in the data frame.</exception>
+    public static ReadOnlySeries GetRequiredNamedSeries(this ReadOnlyDataFrame dataFrame, string name)
+    {
+        ArgumentNullException.ThrowIfNull(dataFrame);
+        ArgumentException.ThrowIfNullOrWhiteSpace(name);
+
+        return dataFrame[name] ?? throw new KeyNotFoundException($"The series '{name}' was not found in the data frame.");
+    }
+
+    /// <summary>
+    /// Gets the series with the specified name and ensures it has elements of type <typeparamref name="T"/>.
+    /// </summary>
+    /// <exception cref="KeyNotFoundException">Thrown when the series with the specified name does not exist in the data frame.</exception>
     /// <exception cref="InvalidDataException">The series does not have elements of type <typeparamref name="T"/></exception>
-    public static ReadOnlySeries<T> Expect<T>(this ReadOnlyDataFrame dataFrame, string name)
+    public static ReadOnlySeries GetRequiredNamedSeriesWithElementType<T>(this ReadOnlyDataFrame dataFrame, string name)
         where T : IEquatable<T>
     {
         ArgumentNullException.ThrowIfNull(dataFrame);
         ArgumentException.ThrowIfNullOrWhiteSpace(name);
 
-        var series = dataFrame[name] ?? throw new KeyNotFoundException($"The series '{name}' was not found in the data frame.");
-        return series.ExpectElementOfType<T>();
+        return dataFrame.GetRequiredNamedSeries(name).CastToSeriesWithElementType<T>();
     }
 }
