+++ test/DSE.Open.Numerics.Tests/ReadOnlyDataFrameExtensionsTests.cs
@@ -6,7 +6,7 @@ namespace DSE.Open.Numerics;
 public sealed class ReadOnlyDataFrameExtensionsTests
 {
     [Fact]
-    public void Expect_WithValidSeries_ShouldReturnTypedSeries()
+    public void GetRequiredNamedSeriesWithElementType_WithValidSeries_ShouldReturnTypedSeries()
     {
         // Arrange
         const string seriesName = "Test Series";
@@ -14,34 +14,34 @@ public sealed class ReadOnlyDataFrameExtensionsTests
         var dataFrame = ReadOnlyDataFrame.Create([expected]);
 
         // Act
-        var series = dataFrame.Expect<int>(seriesName);
+        var series = dataFrame.GetRequiredNamedSeriesWithElementType<int>(seriesName);
 
         // Assert
         Assert.Same(expected, series);
     }
 
     [Fact]
-    public void Expect_WithMissingSeries_ShouldThrowKeyNotFoundException()
+    public void GetRequiredNamedSeriesWithElementType_WithMissingSeries_ShouldThrowKeyNotFoundException()
     {
         // Arrange
         var dataFrame = ReadOnlyDataFrame.Create([]);
 
         // Act
-        void Act() => dataFrame.Expect<int>("Missing Series");
+        void Act() => dataFrame.GetRequiredNamedSeriesWithElementType<int>("Missing Series");
 
         // Assert
         _ = Assert.Throws<KeyNotFoundException>(Act);
     }
 
     [Fact]
-    public void Expect_WithMismatchedType_ShouldThrowInvalidDataException()
+    public void GetRequiredNamedSeriesWithElementType_WithMismatchedType_ShouldThrowInvalidDataException()
     {
         // Arrange
         const string seriesName = "Test Series";
         var dataFrame = ReadOnlyDataFrame.Create([ReadOnlySeries.Create([1, 2, 3], seriesName)]);
 
         // Act
-        void Act() => dataFrame.Expect<string>(seriesName);
+        void Act() => dataFrame.GetRequiredNamedSeriesWithElementType<string>(seriesName);
 
         // Assert
         _ = Assert.Throws<InvalidDataException>(Act);
